bot/requirements.txt
anthropic>=0.96.0
yfinance>=1.2.0
Flask>=3.1.3
python-dotenv>=1.2.1
requests>=2.32.5
APScheduler>=3.10.4
twilio>=9.0.0
pytz>=2024.1
pandas>=2.3.3
numpy>=2.0.2
gunicorn>=22.0.0
